+++ CHANGELOG.md
@@ -13,6 +13,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/)
 * nextcloud.py
 * txt.py: add `exception2text()`
 * winrm.py: make `run_cmd()` and `run_ps()` JEA-aware
+* winrm.py: make `run_cmd()` and `run_ps()` Kerberos-aware
 
 
 ### Fixed ("fix")